interfaces,daemon: move JSON types to the daemon
This patch moves JSON marshaling code from the interface repository to
the daemon API layer. This is the first step towards cleaning up the
return types so that upcoming interface re-mapping patches can stop
changing deeply connected snap.{Plug,Slot}Info types.

Because JSON marshaling requires instance methods we can no longer use
interfaces.Info.MarshalJSON. Instead the code is inlined in the only
place using that type, namely the getInterfaces function in the daemon.
The duplicated tests are removed since this is already tested with
existing API tests.

While doing this I noticed some duplication and some (perhaps) confusing
names, those will be handled in upcoming patches. I added a TODO for
this.
